Update guard params and return types to use correct typings.
The `never` type is meant to be used to tell typescript that a value will never occur(https://www.typescriptlang.org/docs/handbook/basic-types.html#never). This was being used to to allow passing of a stub `ExecutionContext` object(and TypeScript didn't allow `Partial<ExecutionContext`). Ultimately, an empty object was passed and cast to `ExecutionContext` as none of the methods from `ExecutionContext` are being used. Guard parameters were updated to be passed an instance of `ConfigService` instead of an object literal typed as `never`.