Add global types to tsconfig
The UI library in use by course planner(mark-one) currently depends on `@types/styled-components`, which in turn depends on `@types/react-native` (see https://github.com/DefinitelyTyped/DefinitelyTyped/issues/33311). This is to allow styled-components to be used in native projects. Unfortunately, the types provided by `@types/styled-components` conflict with the types (for example, `DOM`, `FormData` etc.) built into TypeScript. The solution for this is to only import the global types needed for this project, which in this cases is `node` and `mocha`.