Install aliased bleeding edge version from NPM
Pinning `mark-one` to version 0.0.1 (which was previously done in 3bbff40abed4fb44a7abf1bac3b614793c461156) pulls in a placeholder "release" from npm that has a `package.json` file and not much else. In order to pull in the bleeding edge of `mark-one`, the version `develop` should be used instead. This points to an NPM release (rather than the `develop` branch on GitHub).

Co-Authored-By: Jonathan Seitz <jseitz@seas.harvard.edu>