Convert AuthModule into Dynamic Module
This is kind of a bigger change, so I'm going to do my best to run
through what's going on here:

Our original Authentication model was always using the SAML Strategy,
but changing the behavior of the validate function in development. This
meant that it was still calling out to the Harvard Key server, but
discarding whatever came back. That's not what we wanted, but I know I
at least was misunderstanding exactly how Passport works.

After banging around in the NestJS and Passport docs for a bit, I
realized that our Authentication process really needs to be completely
different in development. Fortunately, Passport allows you to easily
change the Strategy it uses to authenticate, and NestJS loves Dependency
Injection, so this should be easy, right?

Not quite. The way that the Nest AuthGuard works, the strategy isn't
_really_ a dependency. So I settled on using a Dynamic Module for the
AuthModule, which just means we write it asa function that returns a
module definition. That way it's really injecting a totally new
AuthModule depending on the environment. So in testing, for example, we
can use a TestingStrategy that lets us easily control the outcome of
the authentication to test different scenarios.