Fix integration tests
This implements the new dynamic AuthModule in our integration tests. By
stubbing over the login() method of the the Testing Strategy, we can
easily control which user is logged in, or force it to throw errors.

Along the way, I added the "tests" directory to our tsconfig.json so
that we can get the benefit of the path aliases.