Don't require default Strategy
It looks like the defaultStrategy option was overriding the Strategy
value from config, so we can make that optional and use if it exists, or
use the server value if it doesn't.

I'd say take it out entirely, but I think we may want the ability to
change it for certain tests in the future.