Move jsdom-registration to mocharc
We can use the jsdom-global/register hook to start up the jsdom
environment, instead of doing it manually in the before block.

Also reorganized the require array to match the one in mark-one. Not
sure if it actually matters, but at least it's consistent.

Also removing the "--require ignore-styles" flag from the test script,
as it's being required from the mocharc file.