Re-enable type checking in ts-node
Type checking was previously disabled by setting `TS_NODE_TRANSPILE_ONLY=true` in 36dab2cbe7da33f7460302efabe9df33e8689529. This was previously done because type errors would silently fail the tests for no obvious reason. This appears to no longer be the case, and the absence of type checking caused semi-colons and generic types of be flagged as uncovered.