Create single course with course service
CourseService was previously intended to be able to create multiple `Course` records in the database. However, doing this would require that each course area be checked individually on insert which would have made things tricky to test for no immediately obvious benefit. 

After some discussion, it was decided to update the service to only create single courses for the time being with the idea that this could be refactored later if required