Create View Entities
Working toward the final product of the DTO, I've split all of the
sub-entities into separate database views. Some of these are very small
(e.g., Faculty, Room) but the goal is to essentially cache the query
logic on the database so we don't need to write custom CONCAT statements
in the separate controllers/services.

Originally, I had been looking at using PostgreSQL's materialized views
feature, which caches the data returned from a query, not just the query
itself. However, TypeORM's support for materialized views is not great,
and I don't see a clear path for how to manage refreshing those
materialized views when data is updated.

If the performance of regular views becomes a problem, and support for
materialized views gets better, we may want to look into it as an option
again.