Fix promise resolution for db.init()
The init method on the test database was resolving before the database
was totally ready before, requiring the setInterval call to check on the
status before proceeding. Switching to an explicit resolve statement
in the init method resolves that and makes it easier to use with tests.

Also added docblock comments throughout.