Reorder timezone modifications
Moves the call to utcToZonedTime to between the format and parse calls.

This whole block of code is super-annoying and probably could be
replaced with a regex or db query, in retrospect. But, we may need to do
similar timezone conversions down the line, so good to know that
date-fns-tz does let us do so.