Create typeorm_metadata table in migration
TypeORM automatically creates the typeorm_metadata table in your
database when you generate migrations or sync your schemas, but it
doesn't add code to create the table to the generated migrations.
There's an open PR in the typeorm project that would create the table
when you run the migrations, but imo this should still be done as part
of the migration itself, rather than through a "magic" process.