Add static serve
This module will serve static files located in the `/static` directory,
relative to the location of `app.module.ts`. This directory does not
currently exist, but the compiled client code will be created there when
the app is built for production.

In development, the webpack-dev-middleware registered in the configure
block will override the static server requests, so the missing /static
directory should not be an issue.