Specify Entities for TypeORM
This is the end of an annoying cascade of issues. Since we're bundling
our server code, the glob pattern of `*.entity.ts` was no longer
capturing all of the entities. I briefly played with taking out webpack
bundling altogether and just using tsc, but then we lose out on some of
the webpack plugin features (ignoring packages, replacing tsconfig
paths). This was the best comporomise I could find: individually
importing and listing the Entities one-by-one.

Note that future entities will also need to be added here as well.