Set up tsc to output server code
I hate this less than the previous solution. What we're doing now is
using tsc to transpile all of the server and common code into the
build directory. This required creating a tsconfig.server file, which
extends our default config by ignoring tests and some of the
dev-specific files. We then need to require tsconfig-paths/register in
our node command in the Dockerfile, which required moving that package
to dependencies (instead of devDependencies). But to make that work we
need yet another tsconfig file, which only specifies the
compilerOptions.paths and baseUrl. That file is copied into our final
container as tsconfig.json. Finally, since our directory structure and
file extensions have changed, we need to add a ternary condition to the
entities field in our db configs.