Update type of academicYear and calendarYear to string
When typeorm transforms the type numeric from PostgreSQL to JavaScript, the value is represented as a string (unlike the type integer, which is transformed into a number). This commi updates academicYear and calendarYear type to string to reflect their actual type. For example, this prevents the compiler from allowing a number to be concatenated to a string instead of added. Also, strict comparisons will fail if one side is a number and the other is a string, whereas the compiler will not complain if it falsely thinks the value is a number.