Pre-create faculty areas
The `area` property on the `Faculty` entity does not currently cascade when inserting data. This was causing the `area` property to be `null` in tests, meaning that the area name wasn't applying correctly to the sort order.

In addition to this, test cases for first and last name were changed to put the faculty members in the same area to avoid accidentally testing for area sorting in the wrong test case