Split CourseInstanceList
The full component was starting to get a little unwieldy, so I split the
table itself out into a separate `courseInstanceTable` and all of the
logic into `CoursesPage`.

Similarly, I realized that the logic for defining custom views and then
filtering the columns was going to be very tedious to implement with all
of the column headings and data values hard-coded as separate
`TableHeadingCells` and `TableCells`. So I split the logic for rendering
each individual column off into the `tableFields` file, which is an
array of objects that include necessary metadata alongside a function to
return the proper representation of the data.