Rework table rendering logic
This is primarily focused on getting the header to work correctly,
which requires three separate layers of TableRow inside the TableHead.
There still need to be some finessing of the display for the top row, as
we'll need to define some "invisible" TableHeadingCells to use alongside
the Spring/Fall segments.