Remove ConfigService from FacultyScheduleService and check for invalid empty acadYears array
ConfigService is no longer being used, because if no arguments are provided for the FacultySchedule controller, data for all valid years in the database will be fetched. I am checking that acadYears does not have a length of 0 (this is also being checked in the controller) to avoid the case where the user only provided invalid years in order to prevent an SQL error.