Move 'details' into view options
There's a weird bug in the existing course planner regarding the
"Details" column, and it's unclear if it's supposed to be shown in every
view or if user's should be able to remove it. I've created/edited a
couple different views and sometimes the details column shows up and
sometimes it doesn't, though there isn't an option in the interface to
select it or not.

For right now, I'm going to move this into the list of columns, and as
we progress with that feature we can discuss further what the correct UX
for that column should be.