replace isSEAS as boolean example
Previously, the 'isSEAS' field was being used as an example of a boolean
field. Since that's now an enum, we should use isUndergraduate instead.

Also includes a change made in 07bcbfb on a separate branch, which
replaces offered with actualEnrollment as an example of a semester-level
field. Even though we're not using the translation function for OFFERED
here, there is some overlap in the values between OFFERED and IS_SEAS,
so the check `rawValue in IS_SEAS` is catching those values and passing
them into the isSEASEnumToString function. This doesn't currently cause
a problem, but it's something to watch out for...