Update typings to reflect updated DTO and check that the instance is not null or undefined
We cannot map on faculty unless instance is defined. Originally, we had an array of instances instead of just 'instance', so we did not need to check whether instance was defined. If we don't check that instance is null/undefined and we try to map on faculty, we get the error that we cannot find faculty of undefined