Add createFaculty function to make a POST request for Create Faculty modal
Add a createFaculty function to make a POST request for the Create Faculty modal. Integrate this function into the modal logic and add validation logic for HUID, first name, and last name (The user must provide either a first name or a last name at the minimum to submit the form).