Add tests for the Faculty Admin POST API requests
Check the results of the POST request when it succeeds and when it fails. Make sure that when it succeeds, the correct function was called, that the request was made to the correct URL, and that the result returned is what is expected. Make sure that when the request fails that we are getting the expected error returned.