Updating how sinon stubs are reset
The property names of mockFacultyRepository, mockAreaRepository, etc. are repeated (both have a save method) - Only the first property called save, which happened to be of the mockAreaRepository , was being reset but not the save property of the mockFacultyRepository. Updated the way that sinon stub was being reset so that both save methods are actually reset