Export the users (getCurrentUser) method as part of an object so that we can stub them as part of the TypeScript 3.9.2+ update
It turns out that you can no longer stub free-standing functions in Typescript 3.9.2+ according to this issue:

https://github.com/microsoft/TypeScript/issues/38568
https://github.com/sinonjs/sinon/issues/562

The test in App.test.tsx was failing when it tried to get the callCount for a stub for a free-standing function getCurrentUser, and currently, my package-lock.json has 3.9.5, so that appears to be why. My fix was to use an object where the functions become methods instead.