Proper build arg handling
Upon further review, the environment variables in the build context
aren't actually passed to the docker daemon when building the image.
Instead, we need to define build ARGs for them, and then specify values
in the compose file or as the --build-args argument to docker build.

Fortunately, changing this to use the ARG OPEN_PORT with a default value
of 3000 means that 1.) we don't actually need to change our travis
config, and 2.) we can just assign that ARG to the correct value in
docker-compose and don't need to pass in two arguments.