Limit copies into docker image
Under our new configuration, we don't actually need the client code
inside our container so we can split the build command into two separate
commands and only run build:server in the container. Handling for the
client code still needs to be determined.

I've also added `index.js` to the CMD argument to make it more explicit
which file is being run.

Finally, I've added a .dockerignore file to keep unneeded files out of
the container image. This includes many of the build artifacts, along
with config files that aren't directly needed for running the project.