Update the faculty controller POST action so that area must exist to make the request. Fix tests too
The user should not be able to make a new area when they create a new faculty member. Update the POST action for faculty so that an error is thrown when they try to make a new area while creating a new faculty, and fix the associated tests to reflect this change.