Set SERVER_URL with webpack.definePlugin
Since we won't be able to set or access environment variables at runtime
(ie, in the user browser), we'll need to substitute the SERVER_URL value
at build time with webpack.

The definePlugin does this, but handles string literals a little
strangely, so we need to do the JSON.stringify() in the webpackfile(s)
and ignore the linter/compiler rules in the client code to let us do it.