User SERVER_URL for SAML auth
I'm actually not 100% sure this is correct, but with the change to
running client and server from seprate URLs, I presume we'll need to
give the URL for the server for our SAML request to HarvarKey. Hopefully
we'll be able to test this out soon, and can change if necessary.