switch replacement to process.env.SERVER_URL
I had originally chosen not to use process.env.SERVER_URL as our
replacement value to avoid any confusion about where the value was
coming from. But just using __SERVER_URL__ caused a lot of problems with
our test suite since that value is undefined and defining it would have
required far too much conditional logic that would have been injected
into our production code. So I'm opting for the lesser of the evils and
making our replacement string "process.env.SERVER_URL" instead.

Since webpack does this statically at compile time, it's replacing the
literal string "process.env.SERVER_URL" where it appears in code; it's
not creating a process object with an embedded env object where
SERVER_URL is assigned a value.

But when we run mocha, it's running in a node context where process.env
already exists, so we need to set the process.env.SERVER_URL value in
our .mochainit.ts so that the requests are made correctly.

For testing and development, this should pretty much work as you would
expect it to, so long as the SERVER_URL value is set in .env. When
building the client code for production, you'll need to explicitly set
the variable SERVER_URL to the url for the server instance in the build
environment, i.e.:

    $ SERVER_URL=https://api.seas.harvard.edu/course-planner webpack

If building on Travis, this value will need to be set from the travis
dashboard.