Fix tests broken by change to axios
Our tests were importing axios from the npm package and stubbing out the
get method. This was no longer working once the api methods were changed
to use the single instance of axios defined in request.ts. To resolve:

1. Changed the api test files to import from '../request' instead.
2. Changed the component files to stub out the individual api methods
instead of stubbing out axios's get method.

This second change was made to improve the isolation of the component
tests themselves -- since our components are using the API methods, we
shouldn't care what underlying library they're using to get that data,
only that it returns the data as expected.