Move area colors into constants dir
Our existing way of accessing the area colors as direct values of the
theme had a couple of issues:

1. It relied on the color values being defined in mark-one, which would
make updating/adding colors more difficult
2. It used dynamic properties, which are open to type errors
3. It required string transformations and property checking in the
component body, which can get overly verbose and harder to update.

Instead, we're creating a constant that associates all of the area names
with their color values, then providing a funciton that safely returns
a string that either contains the hex value, or contains nothing.