Fix return value of stub
The API functions were changed to return the user data directly. This
was breaking the test, as the stub was returning an AxiosResponse that
contained teh user data.

Also exports the `ColdApp` component as the default and uses that in the
test to get rid of errors related to the React Hot Loader.