Move User entity into /common/classes
The User entity we defined was not actually connected to any database
model, so there's no reason it needed to be in the server side of the
app. Additionally, the getters for fullName and displayName were not
being included in the API responses, so we'll need to instantiate new
User's on the client side as well.