Create User Controller
Adds the getCurrentUser route that we've been pretending to call in the
client code, along with login and logout routes. The login route simply
throws up an Authentication Guard, then returns the validated user
stored in session to path defined in a query parameter. The logout route
deletes the session, then redirects the user to the official Harvard Key
logout page.

We're also changing the unset behavior in the nest session config, so
that it will remove the session from Redis when deleted