Handle user login flow
Rather than just displaying an error when the user can't log in, we're
catching the specific 401 error returned from our current user route,
then redirecting the client to our login route with a query parameter
indicating the page we're starting from. After authenticating, the user
will be returned to this starting page and will kick off another request
for the current user, which will now be available in the session.