Get currentUser before any other AJAX calls
We need to have the user log in before they can make any other API calls
in the app -- otherwise those calls will hit passports redirect and
fail. This combines the getCurrentUser and getMetadata calls into a
single useEffect block, and only calls for the metadata after the user
has already logged in.