Manually set user in session
In theory, Passport should serialize the user to the session on its own,
but for some reason that is not working correctly in the nest
configuration. So instead, we're just going to manually set the value in
Session. We still need to return the user value though, because
otherwise authentication fails.