Add COOKIE_DOMAIN env variable
Because we're now running the client and server on separate subdomains,
we need to include a "domain" property in the header the specifies a
common domain to allow for cookie sharing between them. I considered
adding a function to parse the two domains and determine a common value
to use, but that adds a lot more complication for the hypothetical cases
where there is no common domain, so it's easier to just use an
environment variable.