Remove waitForElement from findByText calls
FindByText returns a promise on its own, so it should not be wrapped with waitForElement. I was getting the error message: "You seem to have overlapping act() calls, this is not supported. Be sure to await previous act() calls before making a new one." FindByText and waitForElement were both trying to query the dom at once.