Move HarvardKey logic into Auth Controller
Handling the Harvard Key authentication logic through a Passport
Strategy was failing on our QA infrastructure (it seemed to be related
to redirection command). Rather that continue bashing my head into the
wall, I pulled the authentication code directly into the controller,
which now seems to function correctly in both local and QA environments.

I have left in the HarvardKeyStrategy file, which is now just
implementing the passport-custom strategy and checking to see if there's
a user stored in session or else throwing an error.

As an exercise for future development, we may want to re-extract this
authentication code into a standalone nest module that could be imported
into future apps, which could, in turn, export its own @Guard so we'd no
longer need the passport module at all.