Add a basic health check endpoint
This is a _very_ simplistic implementation, which only needed because
our AWS Load Balancer needs some endpoint it can check on a regular
basis to ensure that the container is still up. Rather than use one of
our api endpoints (which would ping our database), we can just return a
200 reponse from an arbitrary endpoint as long as the container is
running.

The nest docs do include a section for setting up more robust health
checks using terminus, but that's out of scope for the current ticket.

https://docs.nestjs.com/recipes/terminus