Add GROUPER_PREFIX environment variable
Presently, Harvard Key is returning a list of all groups for which the
user is a member. We don't need all this information and don't want to
expose it inadvertently, so we can add a GROUPER_PREFIX value and use
that to filter the list.

If our Harvard Key setting can be changed to only include the groups
related to the app, we can remove this.