Move data and utils directory back into the common/__tests__/ directory
'data' and 'utils' directories should not have been removed from the common directory. The code was intentionally organized this way so that when we are bundling the code, we can include any files that do not have '__tests__' in its path. I also reverted the tsconfig.json paths for testData and test-utils.