Restore named exports in API modules
I have not run into the issues with stubbing that prompted the removal
of named exports, but removing those named exports did break some of my
tests that relied on them. Thus, I've restored the named exports of the
inidividual API functions, alongside the named export that includes all
the api functions in a single object. This should provide optimum
compatibility.