Disable secure; set SameSite strict in cookie
Previously I had set "SameSite: None" and "Secure: false", which worked
but dropped warnings in the browser about how that was a bad idea.

I then tried enabling only "Secure", but that was preventing the cookie
from being transmitted back and forth (possibly because of something in
the ELB?), resulting an infinite loop of requests because it couldn't
access the session. I had the same result with all three options
("None", "Strict", and "Lax") for the SameSite option.

Finally, I disabled "Secure" and set SameSite to "Strict", and that
seemed to do the trick to send the session cookie back and forth between
the client and server.