Add course service method findCourses for courses GET request and fix associated tests
The findCourses method was added to sort the course results using TypeORM. We have decided to do a GET request on successful modal submission (when creating or editing a course) instead of sorting on the client side, so having the GET response sorted is necessary. I also adjusted the controller and service tests accordingly to account for these changes.