Add integration tests to check the sorting of courses in the course admin table
These tests check to make sure that the courses are sorted by area in ascending alphabetical order as well as the catalog number by ascending alphabetical order. The test that checks the ordering of the catalog numbers assigns both courses with the same area as a control so that we are only checking the effects of the catalog number.