Derive cookie domain/path from SERVER_URL
The original implementation used 'seas.harvard.edu' becuase I had
mistakenly thought that the cookie needed to be value on both the client
and server domains. This is not the case, and providing only the server
domain is sufficient.

However, our eventual goal is to run all of our server apps out of this
computingapps domain, so to prevent this from being an issue again in
the future we need to provide a path and a domain. So rather than
provide an additional env var for COOKIE_PATH, I figured it would be
easier to use the WHATWG URL API to extract and normalize those values
from the SERVER_URL variable. This has the added benefit of obviating
the COOKIE_DOMAIN variable, so this branch will not require any
.env changes.