Use derived values throughout auth controller
I had so much fun playing with the WHATWG URL API in 618fdec that I
decided to use it everywhere!

OK, so that's only partly true. But while dealing with the cookie
business I did realize that I could use it to resolve a lot of
inconsistent URL normalization logic and move more of the configuration
related things into the ConfigService.

This has the added benefit of making the app more resilient to weird URL
values, with the notable exception that you now cannot run the app if
you don't have SERVER_URL, CLIENT_URL, and CAS_URL set to real URLs in
the environment.