Add logLevel getter to ConfigService
This uses the NODE_ENV value to determine which of the LOG_LEVEL
constants should be used for the current environment.

We may decide to make LOG_LEVEL its own environment variable later on,
as that would let us easily change the logging for a particular
environment without needing to update code -- worth more exploration and
discussion.