Remove the required asterisk from the Undergraduate checkbox
We discussed during the course planning sprint meeting today that having the asterisk next to this checkbox is confusing for the user, because the undergraduate checkbox can be left unchecked in cases where the course is a graduate course. Having the asterisk here implies that the box needs to be checked before continuing.