Fix tests to account for area not needing to exist to create courses
Users are able to create new areas when creating courses, so I removed the tests that check that an exception is thrown when the area does not already exist. In the course entity, the "cascade: ['insert']" option was added for the area so that if the area does not exist at the time of course creation, it will be created.