Update customRender function by setting the test metadata as a fixed value
Discussed this with J Seitz today - AppBlueprint was adding too much complexity to customRender function. Instead of calling useState in the AppBlueprint, we decided it would be better to pass in the entire MetadataContext object as an argument for our testing purposes. Since this is just the test renderer, the state doesn't need to be fully functional. As a result of setting the metadata in this way, we could remove the third argument of metadata from our tests using customRender.