Create log middleware for http requests
This is a small bit of middleware that just instantiates the morgan http
logger, using `winston.http()` method to actually write the logs.

In order to integrate morgan into winston, we needed to provide a stream
wrapper around `winston.http()`, which is just an object that exposes a
`write` method. Since morgan is a function that returns a middleware
handler, our `use()` method just needs to call that handler with the
req, res, and next arguments from express.