Validate log level in Config Service
This adds the log levels back as constants, and brings back a config
service method for getting the log level. This is more of a midpoint
between the two implementations so far (basing the log level on the
node environment and blindly taking the logLevel) in that it still
uses the environment variable direclty, but ensures that it is a valid
logging level before returning it, and provides a fallback of 'error'.