Fix conditionals in .log() function
As in the error() function, the conditional types were not working as
expected so we're switching to an overload.

I also discovered that the way Nest calls `.log()` is not exactly as
expected, as it will pass in a second argument indicating with process
in Nest is generating the message. This required some additional
tweaking to the conditionals.

This is still a very wonky function, but ideally we'll never need to
call this directly and it will only be used by NestJS/TypeORM