Add unit tests for LogService
Logging is a weird thing to test, but this is at least making sure that
the methods in the LogService are passing things through to winston
correctly.

I'll need to do some more hacking to figure out the best way to write
integration tests for things like the LogInterceptor. I'm thinking it
will require stubbing out the `console` object and see what's written
there...