Create Schedule{Block,Entry}View
This adds two database views for the ScheduleBlock, which represents the
blocks of time in the schedule, and the ScheduleEntry, which represents
the individual courses inside those blocks. The latter includes some
additional data that will be shown on hover/click.

Since we'll need a unique "key" value for each block, as well as a way
to join these two entities, we're creating a unique id by concatenating
the prefix, weekday, starttime, endtime, term, and year, then including
that in both views.