Display Log Out only when logged in
This adds a currentUser prop to the AppHeader, and based on its presence
or absence will show or not show the log out link.

Since the AppHeader lives just below the App component, it's easier to
accept the currentUser as a prop, rather than fetching it from context.
We'll need the whole user when we start changing the navigation options
in response to the user's permissions.