Add service method for schedule data
Handles the joining of the schedule blocks and entries, and orders the
blocks so that they can flow from earliest to latest from Monday to
Friday.

I was pleasantly surprised to see that Postgres is ordering the weekday
ENUM in day-of-the-week order, not alphabetic order, presumably because
that's the order was used when defining the enum.