Refactor ScheduleDTO
The original schedule DTO had a lot of nested properties that didn't
necessarily need to be nested and would have made it a little harder to
sort and join things. Those were broken out into individual properties.
The individual course data was also embedded into the top level DTO, and
that was broken into a separate class to clean things up a little.