Add integration tests for Schedule generation
This suite of tests uses the real controller and service methods on top
of our fake database to ensure that the schedule data generated by the
app is in the correct format, and in the corect order.

The ordering tests are particularly tedious, as I ended up using a
series of nested maps to group all of the schedule blocks based on the
different orderables (e.g. weekday, startTime) and then iterated through
the maps to ensure that each met the correct criteria. I chose nested
Maps over nested objects as 1) Maps are directly iterable, and 2) Maps
preserve the insertion order. Because of all the looping I had to do, I
made sure to include some additional tests just to ensure that any tests
were run within a block, and added some additional fake course/faculty
data to provide overlap within some categories.