Remove startTime/endTime from ScheduleEntryView
The time is already captured in the parent "Block", so it's not actually
necessary to include them in each of the entries. These had been
included as a sort of pre-optimization to avoid needing to refomat the
hour/minute values on the front-end, but it's unclear if that's
necessary or even actually helpful.

With that change, we have updates to the DTO, the entity, the migration,
and the controller tests.