Add the FacultyAbsenceModal along with tests to check its behavior
The FacultyAbsenceModal includes a dropdown that the user can use to change the absence type for the faculty for that term. It also includes an area for a general error message. On modal open, it sets the focus to the modal's header. The tests check the open behavior and submit behavior to make sure that the dropdown gets populated appropriately and that onSuccess/onClose gets called appropriately.