Create building blocks for Course Schedule
This are, for the most part, straightforward styled components that will
be used to lay out the schedule. The layout leans heavily on CSS Grid,
and leverages the implicit layout features and precise ordering of
courses in the data to handle some of the weirdness of the scheduling.

One slighly oddity is that our DayBlock will end up being laid over our
WeekBlock, with the Week just handling the time rules that are shown
across the schedule and each Day handling the sessions within it. This
can lead to some visual quirks with days overlapping each other at
narrow widths -- this will be need to be investigated and dealt with
further.