Provide number of day columns as a prop
Related to previous commit (45ae723), we can use the length of the days
list to determine how many day columns we'll need, instead of just
always assuming it will be five. This gives us more flexibility to show
just just a limited subset of the days for a paginated view.