Use fetching boolean to show spinner
A minor tweak on the previous implementation, but we're now using a
dedicated state value to determine whether data is fetching, instead of
just using the length of the schedule prop. This could have been a
problem in cases where the API endpoint returns an empty list, though
that's very much an edge case.