Add tests for Schedule View and Page
Most of the heavy lifting in this layout is being done by the database
(Filtering/Ordering) or by the CSS for the individual blocks, so there
isn't too much that needs to be tested for these components. I am
running a couple tests on the 24-hour to am/pm time conversions in the
WeekBlock from the ScheduleView tests, since it's easier to set up from
that higher level.

As we make this Schedule interface more robust we'll likely need to add
some tests for filtering functionality, but this should be sufficient
for the initial implementation.