Consolidate loadSchedules function logic into useEffect
The loadSchedules function was duplicating a significant amount of logic that was already included in useEffect. This commit added states that checked whether the faculty schedule data was initially loaded (isInitialized) and whether the data was stale (isStaleData), both of which were used in useEffect. I added isInitialized, because I was running into an infinite loop issue on Faculty tab load (since the object value of currentFacultySchedules was different during each fetch, which re-triggered another fetch). I also replaced onClose with onCancel - originally, the modal was closing and then the absence value in the table updated a few seconds later, which seemed like a bad user experience. I figured it would make more sense for the underlying UI to be updated already before the modal closes to prevent confusion after updating a value and submitting the modal form. Also, instead of calling onClose after the try-catch in the FacultyAbsenceModal, the function closeAbsenceModal is called in the .finally of the useEffect in FacultyPage.