Fit schedule to window
These CSS tweaks should make the schedule render better in different
windows sizes. The changes will:

- Make sure that all the columns fit in the week without overlapping
each other
- Allow the width of the Day columns to vary a little, making the width
smaller overall
- Wrap the courseNumbers so we don't lose characters

This still probably needs work, but I think that this is a reasonable
set of constraints for this stage in the UI. I'd hesitate to tweak
things more before we've implemented some of the additional features, in
case those require further changes to the spacing and sizing.