Fix Faculty Admin tests that failed as a result of adding in the filter fields
Now that we have added another row in the table (an additional table header row) for the filter fields, tests that were checking for the number of table rows and table contents needed to be fixed. I also removed the test involving getComputedStyle, since we decided as a team to remove these.