Condense course admin modal error messages to be in one state instead of several; Remove isValid flag
Instead of having a separate state for each error field, condense the error messages into one state, formErrors. The intermediary object, modalError, was needed because when calling setFormErrors directly within each conditional, it wiped out the other error fields. In order to keep all of the errors so that they all show at once when the submit button is clicked, I called setFormErrors at the very end with modalError.The overall course admin error also needed to be separate, because when it is called in the try-catch of the submit button's onClick, if you call setFormErrors, it clears out the other error messages as well. Also, instead of using the isValid flag, I checked that the values of the modalError object were not equal to an empty string before throwing the validation exception.